+++ example/a.ts
@@ -4,16 +4,17 @@ export interface AuthData {
 	id: number
 }
 
-const app = new Elysia()
-	.derive(({ request }) => {
-		const apiKey = request.headers.get('x-api-key')
-		if (!apiKey) return { auth: null }
+const a = new Elysia().decorate({
+	thing: {
+		a: []
+	}
+})
 
-		return { auth: { id: 1 } satisfies AuthData }
+const b = new Elysia()
+	.decorate({
+		thing: {
+			a: ''
+		}
 	})
-	.onBeforeHandle(({ auth, set }) => {
-		console.log(auth?.id)
-		console.log(set.status)
-	})
-
-app['~Volatile']
+	.use(a)
+	.listen(3000)