+++ example/a.ts
@@ -1,20 +1,2 @@
-import { Elysia } from '../src'
-
-export interface AuthData {
-	id: number
-}
-
-const a = new Elysia().decorate({
-	thing: {
-		a: []
-	}
-})
-
-const b = new Elysia()
-	.decorate({
-		thing: {
-			a: ''
-		}
-	})
-	.use(a)
-	.listen(3000)
+import { Elysia, t } from '../src'
+import { openapi } from '@elysiajs/openapi'