+++ CHANGELOG.md
@@ -3,10 +3,16 @@ Improvement:
 - [#1406](https://github.com/elysiajs/elysia/issues/1406) strictly check for 200 inline status code
 - coerce union status value and return type
 - add `BunHTMLBundleLike` to Elysia inline handler
+- prevent Elysia from being a dependency of itself
+- [#1425](https://github.com/elysiajs/elysia/issues/1425) DELETE doesn't ihnerit derive/resolve type
+
+Change:
+- [#1409](https://github.com/elysiajs/elysia/issues/1409) onTransform now doesn't include type as it isn't validated yet, creating confusion
 
 Bug fix:
 - [#1410](https://github.com/elysiajs/elysia/issues/1410) handle union derive/resolve property
 - probably fix mergeDeep attempted to assign to readonly property
+- derive/resolve inherit type in inline handler
 
 # 1.4.5 - 15 Sep 2025
 Improvement: